requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
cryptography>=41.0.7
pyOpenSSL>=23.3.0
